requirements.txt
aiohttp==3.7.4.post0
alembic==1.6.5
aniso8601==9.0.1
async-timeout==3.0.1
attrs==21.2.0
cachelib==0.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dateparser==1.0.0
Faker==8.8.1
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Flash==2.0.0
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
flask-marshmallow==0.10.0
Flask-Migrate==2.7.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
inflect==5.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==2.15.0
marshmallow-sqlalchemy==0.13.2
matplotlib==3.4.2
multidict==5.1.0
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
python-binance==1.0.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pythonping==1.0.16
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.3.23
text-unidecode==1.3
typing-extensions==3.10.0.0
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.4
websockets==9.0.1
Werkzeug==0.16.0
yarl==1.6.3
